FDB_OLAP_Proiect/4_WEB_Model/P4_requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
pandas>=2.2.2
streamlit>=1.36.0
